requirements.txt
openai>=0.10.0
pyyaml>=5.4
python-dotenv>=0.15.0
